# [quantization] Ensure quantized::add stride matches inputs